requirements.txt
pytesseract==0.3.10
pyautogui==0.9.54
openpyxl==3.1.2
pillow==10.1.0